+++ src/main/java/br/com/adacourse/dto/transacao/SaqueReqDTO.java
@@ -2,8 +2,10 @@ package br.com.adacourse.dto.transacao;
 
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record SaqueReqDTO(
 
         @NotNull(message = "Valor é obrigatório")
-        Double valor
+        BigDecimal valor
 ) {}
