+++ src/main/java/br/com/adacourse/dto/transacao/TransacaoRespResumidoDTO.java
@@ -3,13 +3,14 @@ package br.com.adacourse.dto.transacao;
 import br.com.adacourse.enums.TipoTransacao;
 import br.com.adacourse.models.Transacao;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record TransacaoRespResumidoDTO(
 
         Long id,
         TipoTransacao tipo,
-        Double valor,
+        BigDecimal valor,
         LocalDateTime dataHora
 ) {
