+++ src/main/java/br/com/adacourse/dto/transacao/TransferenciaReqDTO.java
@@ -2,11 +2,13 @@ package br.com.adacourse.dto.transacao;
 
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record TransferenciaReqDTO(
 
         @NotNull
         ContaIdDTO contaDestino,
 
         @NotNull
-        Double valor
+        BigDecimal valor
 ){}
